+++ src/plone.server/plone/server/api/files.py
@@ -1,21 +1,21 @@
 # -*- coding: utf-8 -*-
 from aiohttp.web import StreamResponse
 from os.path import basename
+from plone.server import configure
 from plone.server.api.service import DownloadService
 from plone.server.api.service import TraversableDownloadService
 from plone.server.api.service import TraversableFieldService
 from plone.server.interfaces import IFileManager
-from zope.component import getMultiAdapter
 from plone.server.interfaces import IResource
+from plone.server.interfaces import IStaticFile
+from zope.component import getMultiAdapter
 
 import aiohttp
 import mimetypes
-from plone.server.configure import service
-from plone.server.interfaces import IStaticFile
 
 
 # Static File
-@service(context=IStaticFile, method='GET', permission='plone.AccessContent')
+@configure.service(context=IStaticFile, method='GET', permission='plone.AccessContent')
 class DefaultGET(DownloadService):
     async def __call__(self):
         if hasattr(self.context, '_file_path'):
@@ -34,8 +34,8 @@ class DefaultGET(DownloadService):
 
 
 # Field File
-@service(context=IResource, method='PATCH', permission='plone.ModifyContent',
-         name='@upload')
+@configure.service(context=IResource, method='PATCH', permission='plone.ModifyContent',
+                   name='@upload')
 class UploadFile(TraversableFieldService):
 
     async def __call__(self):
@@ -46,8 +46,8 @@ class UploadFile(TraversableFieldService):
         return await adapter.upload()
 
 
-@service(context=IResource, method='GET', permission='plone.ViewContent',
-         name='@download')
+@configure.service(context=IResource, method='GET', permission='plone.ViewContent',
+                   name='@download')
 class DownloadFile(TraversableDownloadService):
 
     async def __call__(self):
@@ -58,8 +58,8 @@ class DownloadFile(TraversableDownloadService):
         return await adapter.download()
 
 
-@service(context=IResource, method='POST', permission='plone.ModifyContent',
-         name='@tusupload')
+@configure.service(context=IResource, method='POST', permission='plone.ModifyContent',
+                   name='@tusupload')
 class TusCreateFile(UploadFile):
 
     async def __call__(self):
@@ -70,8 +70,8 @@ class TusCreateFile(UploadFile):
         return await adapter.tus_create()
 
 
-@service(context=IResource, method='HEAD', permission='plone.ModifyContent',
-         name='@tusupload')
+@configure.service(context=IResource, method='HEAD', permission='plone.ModifyContent',
+                   name='@tusupload')
 class TusHeadFile(UploadFile):
 
     async def __call__(self):
@@ -82,8 +82,8 @@ class TusHeadFile(UploadFile):
         return await adapter.tus_head()
 
 
-@service(context=IResource, method='PATCH', permission='plone.ModifyContent',
-         name='@tusupload')
+@configure.service(context=IResource, method='PATCH', permission='plone.ModifyContent',
+                   name='@tusupload')
 class TusPatchFile(UploadFile):
 
     async def __call__(self):
@@ -94,8 +94,8 @@ class TusPatchFile(UploadFile):
         return await adapter.tus_patch()
 
 
-@service(context=IResource, method='OPTIONS', permission='plone.AccessPreflight',
-         name='@tusupload')
+@configure.service(context=IResource, method='OPTIONS', permission='plone.AccessPreflight',
+                   name='@tusupload')
 class TusOptionsFile(UploadFile):
 
     async def __call__(self):